Allow compilations to be chained.
The chaining allows subsequent submissions in the REPL to get access to previously declared variables.

This also introduces scoping so that we can lookup variables and type check them properly.

Fixes a couple of issue that were reported in the C# version.

I also took a moment to correct some *tests* failing due to CompilationUnit and EndOfFileToken not being taken into account in the tests; may need to loop back around on this as the video progresses.  The fixes are isolated Parser_UnaryExpression_HonorsPrecedences and Parser_BinaryExpression_HonorsPrecedences.