Allow REPL to force evaluation.
Sometimes our heuristic for detecting whether the submission is complete doesn't work, in which case we want the user to have the ability to force the completion.  We do this by adding two blank lines.

Renamed test to match conventions.

Add ability to get lexer diagnostics.

When parsing tokens only, we currently have no way to get access to the lexer's diagnostics.  This adds an optional out param.

The real fix, of course, would be to mirror what Roslyn is doing, which is exposing the diagnostics on the actual tokens, which makes the output neatly self-contained.

Add support for string literals.